+++ stopDevelopmentEnvironment.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
-DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
-pushd $DIR >/dev/null
-docker-compose -p wenet_interaction_protocol_engine_services_dev -f src/dev/docker/docker-compose.yml down --remove-orphans
-docker stop wenet_interaction_protocol_engine_dev
-docker rm wenet_interaction_protocol_engine_dev
-popd >/dev/null
+if [ -f /.dockerenv ]; then
+   echo "You can not stop the development environment inside a docker container"
+else
+	DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
+	pushd $DIR >/dev/null
+	docker-compose -p wenet_interaction_protocol_engine_services_dev -f src/dev/docker/docker-compose.yml down --remove-orphans
+	docker stop wenet_interaction_protocol_engine_dev
+	docker rm wenet_interaction_protocol_engine_dev
+	popd >/dev/null
+fi
